Simplify transaction to be a part of BlockinfoResponse
