+++ app/src/main/java/io/github/jangerhard/BitcoinWalletTracker/adapter/TransactionAdapter.java
@@ -10,8 +10,8 @@ import android.widget.TextView;
 import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.RecyclerView;
 import io.github.jangerhard.BitcoinWalletTracker.R;
+import io.github.jangerhard.BitcoinWalletTracker.model.BlockinfoResponse.Transaction;
 import io.github.jangerhard.BitcoinWalletTracker.utilities.BitcoinUtils;
-import io.github.jangerhard.BitcoinWalletTracker.model.Transaction;
 
 public class TransactionAdapter extends RecyclerView.Adapter<TransactionAdapter.MyViewHolder> {
